调整代码，整合模型，避免editors import 过多class
