+++ mmagic/models/editors/__init__.py
@@ -9,7 +9,8 @@ from .controlnet import ControlStableDiffusion
 from .cyclegan import CycleGAN
 from .dcgan import DCGAN
 from .ddpm import DenoisingUnet
-from .deblurganv2 import (DeblurGanV2, FPNInception, FPNMobileNet, DoubleGan)
+from .deblurganv2 import (DeblurGanV2, DeblurGanV2Generator,
+                          DeblurGanV2Discriminator)
 from .deepfillv1 import (ContextualAttentionModule, ContextualAttentionNeck,
                          DeepFillDecoder, DeepFillEncoder, DeepFillRefiner,
                          DeepFillv1Discriminators, DeepFillv1Inpaintor)
@@ -89,6 +90,6 @@ __all__ = [
     'StyleGAN3Generator', 'InstColorization', 'NAFBaseline',
     'NAFBaselineLocal', 'NAFNet', 'NAFNetLocal', 'DenoisingUnet',
     'ClipWrapper', 'EG3D', 'Restormer', 'SwinIRNet', 'StableDiffusion',
-    'ControlStableDiffusion', 'DreamBooth', 'DeblurGanV2', 'FPNInception',
-    'FPNMobileNet', 'DoubleGan'
+    'ControlStableDiffusion', 'DreamBooth', 'DeblurGanV2', 'DeblurGanV2Generator',
+    'DeblurGanV2Discriminator'
 ]
