+++ mmagic/models/editors/deblurganv2/__init__.py
@@ -1,11 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .deblurganv2 import DeblurGanV2
-from .deblurganv2_fpn_inception_generators import FPNInception
-from .deblurganv2_fpn_dense_generators import FPNDense
-from .deblurganv2_fpn_inception_simple_generators import FPNInceptionSimple
-from .deblurganv2_fpn_mobilenet_generators import FPNMobileNet
-from .deblurganv2_unet_seresnext import UNetSEResNext
-from .deblurganv2_fpn_inception_discriminator import DoubleGan
+from .deblurganv2_fpn_inception_generators import DeblurGanV2Generator
+from .deblurganv2_fpn_inception_discriminator import DeblurGanV2Discriminator
 
-__all__ = ['DeblurGanV2', 'DoubleGan', 'FPNInception', 'FPNDense',
-           'FPNInceptionSimple', 'FPNMobileNet', 'UNetSEResNext']
+__all__ = ['DeblurGanV2', 'DeblurGanV2Generator', 'DeblurGanV2Discriminator']
