--- mmagic/models/editors/deblurganv2/deblurganv2_fpn_mobilenet_generators.py
@@ -1,279 +1,0 @@
-import math
-
-import torch
-import torch.nn as nn
-
-from mmagic.models.editors.deblurganv2.deblurganv2_util import get_norm_layer
-from mmagic.registry import MODELS
-
-
-def conv_bn(inp, oup, stride):
-    return nn.Sequential(
-        nn.Conv2d(inp, oup, 3, stride, 1, bias=False),
-        nn.BatchNorm2d(oup),
-        nn.ReLU6(inplace=True)
-    )
-
-
-def conv_1x1_bn(inp, oup):
-    return nn.Sequential(
-        nn.Conv2d(inp, oup, 1, 1, 0, bias=False),
-        nn.BatchNorm2d(oup),
-        nn.ReLU6(inplace=True)
-    )
-
-
-class InvertedResidual(nn.Module):
-    def __init__(self, inp, oup, stride, expand_ratio):
-        super(InvertedResidual, self).__init__()
-        self.stride = stride
-        assert stride in [1, 2]
-
-        hidden_dim = round(inp * expand_ratio)
-        self.use_res_connect = self.stride == 1 and inp == oup
-
-        if expand_ratio == 1:
-            self.conv = nn.Sequential(
-                # dw
-                nn.Conv2d(hidden_dim, hidden_dim, 3, stride, 1, groups=hidden_dim, bias=False),
-                nn.BatchNorm2d(hidden_dim),
-                nn.ReLU6(inplace=True),
-                # pw-linear
-                nn.Conv2d(hidden_dim, oup, 1, 1, 0, bias=False),
-                nn.BatchNorm2d(oup),
-            )
-        else:
-            self.conv = nn.Sequential(
-                # pw
-                nn.Conv2d(inp, hidden_dim, 1, 1, 0, bias=False),
-                nn.BatchNorm2d(hidden_dim),
-                nn.ReLU6(inplace=True),
-                # dw
-                nn.Conv2d(hidden_dim, hidden_dim, 3, stride, 1, groups=hidden_dim, bias=False),
-                nn.BatchNorm2d(hidden_dim),
-                nn.ReLU6(inplace=True),
-                # pw-linear
-                nn.Conv2d(hidden_dim, oup, 1, 1, 0, bias=False),
-                nn.BatchNorm2d(oup),
-            )
-
-    def forward(self, x):
-        if self.use_res_connect:
-            return x + self.conv(x)
-        else:
-            return self.conv(x)
-
-
-class MobileNetV2(nn.Module):
-    def __init__(self, n_class=1000, input_size=224, width_mult=1.):
-        super(MobileNetV2, self).__init__()
-        block = InvertedResidual
-        input_channel = 32
-        last_channel = 1280
-        interverted_residual_setting = [
-            # t, c, n, s
-            [1, 16, 1, 1],
-            [6, 24, 2, 2],
-            [6, 32, 3, 2],
-            [6, 64, 4, 2],
-            [6, 96, 3, 1],
-            [6, 160, 3, 2],
-            [6, 320, 1, 1],
-        ]
-
-        # building first layer
-        assert input_size % 32 == 0
-        input_channel = int(input_channel * width_mult)
-        self.last_channel = int(last_channel * width_mult) if width_mult > 1.0 else last_channel
-        self.features = [conv_bn(3, input_channel, 2)]
-        # building inverted residual blocks
-        for t, c, n, s in interverted_residual_setting:
-            output_channel = int(c * width_mult)
-            for i in range(n):
-                if i == 0:
-                    self.features.append(block(input_channel, output_channel, s, expand_ratio=t))
-                else:
-                    self.features.append(block(input_channel, output_channel, 1, expand_ratio=t))
-                input_channel = output_channel
-        # building last several layers
-        self.features.append(conv_1x1_bn(input_channel, self.last_channel))
-        # make it nn.Sequential
-        self.features = nn.Sequential(*self.features)
-
-        # building classifier
-        self.classifier = nn.Sequential(
-            nn.Dropout(0.2),
-            nn.Linear(self.last_channel, n_class),
-        )
-
-        self._initialize_weights()
-
-    def forward(self, x):
-        x = self.features(x)
-        x = x.mean(3).mean(2)
-        x = self.classifier(x)
-        return x
-
-    def _initialize_weights(self):
-        for m in self.modules():
-            if isinstance(m, nn.Conv2d):
-                n = m.kernel_size[0] * m.kernel_size[1] * m.out_channels
-                m.weight.data.normal_(0, math.sqrt(2. / n))
-                if m.bias is not None:
-                    m.bias.data.zero_()
-            elif isinstance(m, nn.BatchNorm2d):
-                m.weight.data.fill_(1)
-                m.bias.data.zero_()
-            elif isinstance(m, nn.Linear):
-                n = m.weight.size(1)
-                m.weight.data.normal_(0, 0.01)
-                m.bias.data.zero_()
-
-
-class FPNHead(nn.Module):
-    def __init__(self, num_in, num_mid, num_out):
-        super().__init__()
-
-        self.block0 = nn.Conv2d(num_in, num_mid, kernel_size=3, padding=1, bias=False)
-        self.block1 = nn.Conv2d(num_mid, num_out, kernel_size=3, padding=1, bias=False)
-
-    def forward(self, x):
-        x = nn.functional.relu(self.block0(x), inplace=True)
-        x = nn.functional.relu(self.block1(x), inplace=True)
-        return x
-
-
-@MODELS.register_module()
-class FPNMobileNet(nn.Module):
-
-    def __init__(self, norm_layer, output_ch=3, num_filters=64, num_filters_fpn=128, pretrained=None):
-        super().__init__()
-
-        # Feature Pyramid Network (FPN) with four feature maps of resolutions
-        # 1/4, 1/8, 1/16, 1/32 and `num_filters` filters for all feature maps.
-
-        self.fpn = FPN(num_filters=num_filters_fpn, norm_layer=get_norm_layer(norm_type=norm_layer), pretrained=pretrained)
-
-        # The segmentation heads on top of the FPN
-
-        self.head1 = FPNHead(num_filters_fpn, num_filters, num_filters)
-        self.head2 = FPNHead(num_filters_fpn, num_filters, num_filters)
-        self.head3 = FPNHead(num_filters_fpn, num_filters, num_filters)
-        self.head4 = FPNHead(num_filters_fpn, num_filters, num_filters)
-
-        self.smooth = nn.Sequential(
-            nn.Conv2d(4 * num_filters, num_filters, kernel_size=3, padding=1),
-            norm_layer(num_filters),
-            nn.ReLU(),
-        )
-
-        self.smooth2 = nn.Sequential(
-            nn.Conv2d(num_filters, num_filters // 2, kernel_size=3, padding=1),
-            norm_layer(num_filters // 2),
-            nn.ReLU(),
-        )
-
-        self.final = nn.Conv2d(num_filters // 2, output_ch, kernel_size=3, padding=1)
-
-    def unfreeze(self):
-        self.fpn.unfreeze()
-
-    def forward(self, x):
-
-        map0, map1, map2, map3, map4 = self.fpn(x)
-
-        map4 = nn.functional.upsample(self.head4(map4), scale_factor=8, mode="nearest")
-        map3 = nn.functional.upsample(self.head3(map3), scale_factor=4, mode="nearest")
-        map2 = nn.functional.upsample(self.head2(map2), scale_factor=2, mode="nearest")
-        map1 = nn.functional.upsample(self.head1(map1), scale_factor=1, mode="nearest")
-
-        smoothed = self.smooth(torch.cat([map4, map3, map2, map1], dim=1))
-        smoothed = nn.functional.upsample(smoothed, scale_factor=2, mode="nearest")
-        smoothed = self.smooth2(smoothed + map0)
-        smoothed = nn.functional.upsample(smoothed, scale_factor=2, mode="nearest")
-
-        final = self.final(smoothed)
-        res = torch.tanh(final) + x
-
-        return torch.clamp(res, min=-1, max=1)
-
-
-class FPN(nn.Module):
-
-    def __init__(self, norm_layer, num_filters=128, pretrained=None):
-        """Creates an `FPN` instance for feature extraction.
-        Args:
-          num_filters: the number of filters in each output pyramid level
-          pretrained: use ImageNet pre-trained backbone feature extractor
-        """
-
-        super().__init__()
-        net = MobileNetV2(n_class=1000)
-
-        if pretrained:
-            #Load weights into the project directory
-            if torch.cuda.is_available():
-                state_dict = torch.load(pretrained) # add map_location='cpu' if no gpu
-            else:
-                state_dict = torch.load(pretrained, map_location='cpu')  # add map_location='cpu' if no gpu
-            net.load_state_dict(state_dict)
-        self.features = net.features
-
-        self.enc0 = nn.Sequential(*self.features[0:2])
-        self.enc1 = nn.Sequential(*self.features[2:4])
-        self.enc2 = nn.Sequential(*self.features[4:7])
-        self.enc3 = nn.Sequential(*self.features[7:11])
-        self.enc4 = nn.Sequential(*self.features[11:16])
-
-        self.td1 = nn.Sequential(nn.Conv2d(num_filters, num_filters, kernel_size=3, padding=1),
-                                 norm_layer(num_filters),
-                                 nn.ReLU(inplace=True))
-        self.td2 = nn.Sequential(nn.Conv2d(num_filters, num_filters, kernel_size=3, padding=1),
-                                 norm_layer(num_filters),
-                                 nn.ReLU(inplace=True))
-        self.td3 = nn.Sequential(nn.Conv2d(num_filters, num_filters, kernel_size=3, padding=1),
-                                 norm_layer(num_filters),
-                                 nn.ReLU(inplace=True))
-
-        self.lateral4 = nn.Conv2d(160, num_filters, kernel_size=1, bias=False)
-        self.lateral3 = nn.Conv2d(64, num_filters, kernel_size=1, bias=False)
-        self.lateral2 = nn.Conv2d(32, num_filters, kernel_size=1, bias=False)
-        self.lateral1 = nn.Conv2d(24, num_filters, kernel_size=1, bias=False)
-        self.lateral0 = nn.Conv2d(16, num_filters // 2, kernel_size=1, bias=False)
-
-        for param in self.features.parameters():
-            param.requires_grad = False
-
-    def unfreeze(self):
-        for param in self.features.parameters():
-            param.requires_grad = True
-
-
-    def forward(self, x):
-
-        # Bottom-up pathway, from ResNet
-        enc0 = self.enc0(x)
-
-        enc1 = self.enc1(enc0) # 256
-
-        enc2 = self.enc2(enc1) # 512
-
-        enc3 = self.enc3(enc2) # 1024
-
-        enc4 = self.enc4(enc3) # 2048
-
-        # Lateral connections
-
-        lateral4 = self.lateral4(enc4)
-        lateral3 = self.lateral3(enc3)
-        lateral2 = self.lateral2(enc2)
-        lateral1 = self.lateral1(enc1)
-        lateral0 = self.lateral0(enc0)
-
-        # Top-down pathway
-        map4 = lateral4
-        map3 = self.td1(lateral3 + nn.functional.upsample(map4, scale_factor=2, mode="nearest"))
-        map2 = self.td2(lateral2 + nn.functional.upsample(map3, scale_factor=2, mode="nearest"))
-        map1 = self.td3(lateral1 + nn.functional.upsample(map2, scale_factor=2, mode="nearest"))
-        return lateral0, map1, map2, map3, map4
-
