+++ mmagic/models/editors/deblurganv2/deblurganv2_util.py
@@ -713,17 +713,24 @@ class DiscLossWGANGP(DiscLossLS):
     def calc_gradient_penalty(self, netD, real_data, fake_data):
         alpha = torch.rand(1, 1)
         alpha = alpha.expand(real_data.size())
-        alpha = alpha.cuda()
+        if torch.cuda.is_available():
+            alpha = alpha.cuda()
 
         interpolates = alpha * real_data + ((1 - alpha) * fake_data)
 
-        interpolates = interpolates.cuda()
+        if torch.cuda.is_available():
+            interpolates = interpolates.cuda()
         interpolates = Variable(interpolates, requires_grad=True)
 
         disc_interpolates = netD.forward(interpolates)
 
-        gradients = autograd.grad(outputs=disc_interpolates, inputs=interpolates,
-                                  grad_outputs=torch.ones(disc_interpolates.size()).cuda(),
+        if torch.cuda.is_available():
+            gradients = autograd.grad(outputs=disc_interpolates, inputs=interpolates,
+                                      grad_outputs=torch.ones(disc_interpolates.size()).cuda() ,
+                                      create_graph=True, retain_graph=True, only_inputs=True)[0]
+        else:
+            gradients = autograd.grad(outputs=disc_interpolates, inputs=interpolates,
+                                  grad_outputs=torch.ones(disc_interpolates.size()),
                                   create_graph=True, retain_graph=True, only_inputs=True)[0]
 
         gradient_penalty = ((gradients.norm(2, dim=1) - 1) ** 2).mean() * self.LAMBDA
